requirements.txt
selenium
webdriver-manager
bs4
python-dotenv